Moved service providers under the Providers namespace
