+++ src/Framework/Routing/RouteServiceProvider.php
@@ -3,7 +3,7 @@
 namespace MVPS\Lumis\Framework\Routing;
 
 use Closure;
-use MVPS\Lumis\Framework\Support\ServiceProvider;
+use MVPS\Lumis\Framework\Providers\ServiceProvider;
 
 class RouteServiceProvider extends ServiceProvider
 {
